requirements.txt
pyautogen>=0.2.0
playwright>=1.40.0
python-dotenv>=1.0.0
httpx>=0.25.0
google-generativeai>=0.3.0
